refactor: decompose FolderProcessorApp into UICreationMixin, FileOperationsMixin, ProcessingMixin
Extract 45 methods into 3 mixin files:
- folder_tool_ui.py (717 lines): UI widget creation, icon setup, folder selection
- folder_tool_file_ops.py (1178 lines): File validation, archive extraction, copying, dedup
- folder_tool_processing.py (983 lines): Processing workflow, reports, ZIP creation, dialogs

Reduces Folders_Tool_r0.py from 3,289 to 500 lines (-84.8%)
